+++ packages/typespec-autorest/test/azure-core-types.test.ts
@@ -1,4 +1,4 @@
-import { deepStrictEqual } from "assert";
+import { deepStrictEqual, ok } from "assert";
 import { describe, it } from "vitest";
 import { compileOpenAPI } from "./test-host.js";
 
@@ -68,6 +68,111 @@ describe("armResourceIdentifier", () => {
           },
         ],
       },
+    });
+  });
+});
+
+describe("@uniqueItems", () => {
+  it("defines array with uniqueItems inline", async () => {
+    const res = await compileOpenAPI(
+      `
+     ${base}
+      model Pet { @uniqueItems names: string[] };
+      `,
+      { preset: "azure" },
+    );
+
+    ok(res.definitions);
+    ok(res.definitions.Pet);
+    ok(res.definitions.Pet.properties);
+    ok(res.definitions.Pet.properties.names, "expected definition named names");
+    deepStrictEqual(res.definitions.Pet.properties.names, {
+      type: "array",
+      uniqueItems: true,
+      items: { type: "string" },
+    });
+    deepStrictEqual(res.definitions.Pet.properties.names.uniqueItems, true);
+  });
+
+  it("defines a named array with uniqueItems using model is", async () => {
+    const res = await compileOpenAPI(
+      `
+     ${base}
+      @uniqueItems
+      model PetNames is string[] {}
+      model Pet { names: PetNames };
+      `,
+      { preset: "azure" },
+    );
+    ok(res.definitions);
+    ok(res.definitions.Pet);
+    ok(res.definitions.Pet.properties);
+    ok(res.definitions.Pet.properties.names, "expected definition named names");
+    deepStrictEqual(res.definitions.Pet.properties.names, {
+      $ref: "#/definitions/PetNames",
+      uniqueItems: true,
+    });
+    deepStrictEqual(res.definitions.PetNames, {
+      items: { type: "string" },
+      type: "array",
+      uniqueItems: true,
+    });
+  });
+});
+
+describe("azureLocation", () => {
+  it("defines property with azureLocation type inline", async () => {
+    const res = await compileOpenAPI(
+      `
+     ${base}
+      model Pet { @Azure.ResourceManager.CommonTypes.Private.inlineAzureType location: Azure.Core.azureLocation };
+      `,
+      { preset: "azure" },
+    );
+
+    ok(res.definitions);
+    ok(res.definitions.Pet);
+    ok(res.definitions.Pet.properties);
+    ok(res.definitions.Pet.properties.location, "expected definition named location");
+    deepStrictEqual(res.definitions.Pet.properties.location, {
+      type: "string",
+      description: "Represents an Azure geography region where supported resource providers live.",
+    });
+  });
+  it("defines property with azureLocation type inline but does not override description", async () => {
+    const res = await compileOpenAPI(
+      `
+     ${base}
+      model Pet { /** The azure location */ @Azure.ResourceManager.CommonTypes.Private.inlineAzureType location: Azure.Core.azureLocation };
+      `,
+      { preset: "azure" },
+    );
+
+    ok(res.definitions);
+    ok(res.definitions.Pet);
+    ok(res.definitions.Pet.properties);
+    ok(res.definitions.Pet.properties.location, "expected definition named location");
+    deepStrictEqual(res.definitions.Pet.properties.location, {
+      type: "string",
+      description: "The azure location",
+    });
+  });
+  it("defines property with azureLocation type as ref without decorator", async () => {
+    const res = await compileOpenAPI(
+      `
+     ${base}
+      model Pet { /** The azure location */ location: Azure.Core.azureLocation };
+      `,
+      { preset: "azure" },
+    );
+
+    ok(res.definitions);
+    ok(res.definitions.Pet);
+    ok(res.definitions.Pet.properties);
+    ok(res.definitions.Pet.properties.location, "expected definition named location");
+    deepStrictEqual(res.definitions.Pet.properties.location, {
+      $ref: "#/definitions/Azure.Core.azureLocation",
+      description: "The azure location",
     });
   });
 });