+++ packages/typespec-azure-resource-manager/src/resource.ts
@@ -23,6 +23,7 @@ import {
 import { useStateMap } from "@typespec/compiler/utils";
 import { getHttpOperation, isPathParam } from "@typespec/http";
 import { $autoRoute, getParentResource, getSegment } from "@typespec/rest";
+
 import {
   ArmProviderNameValueDecorator,
   ArmResourceOperationsDecorator,
@@ -37,7 +38,11 @@ import {
   SubscriptionResourceDecorator,
   TenantResourceDecorator,
 } from "../generated-defs/Azure.ResourceManager.js";
-import { CustomAzureResourceDecorator } from "../generated-defs/Azure.ResourceManager.Legacy.js";
+import {
+  ArmExternalTypeDecorator,
+  CustomAzureResourceDecorator,
+  CustomResourceOptions,
+} from "../generated-defs/Azure.ResourceManager.Legacy.js";
 import { reportDiagnostic } from "./lib.js";
 import {
   getArmProviderNamespace,
@@ -79,6 +84,19 @@ export interface ArmResourceDetailsBase {
   /** A reference to the TypeSpec type */
   typespecType: Model;
 }
+
+export const [isArmExternalType, setArmExternalType] = useStateMap<Model, boolean>(
+  ArmStateKeys.armExternalType,
+);
+
+export const $armExternalType: ArmExternalTypeDecorator = (
+  context: DecoratorContext,
+  entity: Model,
+) => {
+  const { program } = context;
+  if (isTemplateDeclaration(entity)) return;
+  setArmExternalType(program, entity, true);
+};
 
 /** Details for RP resources */
 export interface ArmResourceDetails extends ArmResourceDetailsBase {
@@ -188,11 +206,12 @@ export const $armVirtualResource: ArmVirtualResourceDecorator = (
 export const $customAzureResource: CustomAzureResourceDecorator = (
   context: DecoratorContext,
   entity: Model,
+  options?: CustomResourceOptions,
 ) => {
   const { program } = context;
+  const optionsValue = options ?? { isAzureResource: false };
   if (isTemplateDeclaration(entity)) return;
-
-  program.stateMap(ArmStateKeys.customAzureResource).set(entity, "Custom");
+  setCustomResource(program, entity, optionsValue);
 };
 
 function getProperty(
@@ -245,6 +264,12 @@ export function getArmVirtualResourceDetails(
   return undefined;
 }
 
+const [getCustomResourceOptions, setCustomResource] = useStateMap<Model, CustomResourceOptions>(
+  ArmStateKeys.customAzureResource,
+);
+
+export { getCustomResourceOptions };
+
 /**
  * Determine if the given model is a custom resource.
  * @param program The program to process.
@@ -252,7 +277,8 @@ export function getArmVirtualResourceDetails(
  * @returns true if the model or any model it extends is marked as a resource, otherwise false.
  */
 export function isCustomAzureResource(program: Program, target: Model): boolean {
-  if (program.stateMap(ArmStateKeys.customAzureResource).has(target)) return true;
+  const resourceOptions = getCustomResourceOptions(program, target);
+  if (resourceOptions) return true;
   if (target.baseModel) return isCustomAzureResource(program, target.baseModel);
   return false;
 }
@@ -731,7 +757,10 @@ export function getArmResourceInfo(
 export function getArmResourceKind(resourceType: Model): ArmResourceKind | undefined {
   if (resourceType.baseModel) {
     const coreType = resourceType.baseModel;
-    if (coreType.name.startsWith("TrackedResource")) {
+    if (
+      coreType.name.startsWith("TrackedResource") ||
+      coreType.name.startsWith("LegacyTrackedResource")
+    ) {
       return "Tracked";
     } else if (coreType.name.startsWith("ProxyResource")) {
       return "Proxy";
@@ -906,7 +935,7 @@ export const $identifiers: IdentifiersDecorator = (
 };
 
 /**
- * This function returns identifiers using the @identifiers decorator
+ * This function returns identifiers using the '@identifiers' decorator
  *
  * @param program The program to process.
  * @param entity The array model type to check.
@@ -917,7 +946,7 @@ export function getArmIdentifiers(program: Program, entity: ModelProperty): stri
 }
 
 /**
- * This function returns identifiers using the @key decorator.
+ * This function returns identifiers using the '@key' decorator.
  *
  * @param program The program to process.
  * @param entity The array model type to check.