+++ packages/typespec-azure-resource-manager/test/resource-resolution.test.ts
@@ -1049,4 +1049,327 @@ interface GenericResources {
       { operationGroup: "Operations", name: "list", kind: "other" },
     ]);
   });
+
+  it("collects operation information for private endpoints", async () => {
+    const { program, diagnostics } = await compileAndDiagnose(`
+
+using Azure.Core;
+
+/** Contoso Resource Provider management API. */
+@armProviderNamespace
+@service(#{ title: "ContosoProviderHubClient" })
+@versioned(Versions)
+namespace Microsoft.ContosoProviderHub;
+
+/** Contoso API versions */
+enum Versions {
+  /** 2021-10-01-preview version */
+  @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
+  @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v5)
+  v2021_20_01_preview: "2021-10-01-preview",
+}
+
+// For more information about the proxy vs tracked,
+// see https://armwiki.azurewebsites.net/rp_onboarding/tracked_vs_proxy_resources.html?q=proxy%20resource
+/** A ContosoProviderHub resource */
+model Employee is TrackedResource<EmployeeProperties> {
+  ...ResourceNameParameter<Employee>;
+}
+
+/** Employee properties */
+model EmployeeProperties {
+  /** Age of employee */
+  age?: int32;
+
+  /** City of employee */
+  city?: string;
+
+  /** Profile of employee */
+  @encode("base64url")
+  profile?: bytes;
+
+  /** The status of the last operation. */
+  @visibility(Lifecycle.Read)
+  provisioningState?: ProvisioningState;
+}
+
+/** The provisioning state of a resource. */
+@lroStatus
+union ProvisioningState {
+  ResourceProvisioningState,
+
+  /** The resource is being provisioned */
+  Provisioning: "Provisioning",
+
+  /** The resource is updating */
+  Updating: "Updating",
+
+  /** The resource is being deleted */
+  Deleting: "Deleting",
+
+  /** The resource create request has been accepted */
+  Accepted: "Accepted",
+
+  string,
+}
+
+interface Operations extends Azure.ResourceManager.Operations {}
+
+model PrivateEndpointConnection is PrivateEndpointConnectionResource;
+alias PrivateEndpointOperations = PrivateEndpoints<PrivateEndpointConnection>;
+
+@armResourceOperations
+interface Employees {
+  get is ArmResourceRead<Employee>;
+  createOrUpdate is ArmResourceCreateOrReplaceAsync<Employee>;
+  update is ArmCustomPatchSync<
+    Employee,
+    Azure.ResourceManager.Foundations.ResourceUpdateModel<Employee, EmployeeProperties>
+  >;
+  delete is ArmResourceDeleteSync<Employee>;
+  listByResourceGroup is ArmResourceListByParent<Employee>;
+  listBySubscription is ArmListBySubscription<Employee>;
+  /** A sample resource action that move employee to different location */
+  move is ArmResourceActionSync<Employee, MoveRequest, MoveResponse>;
+
+  /** A sample HEAD operation to check resource existence */
+  checkExistence is ArmResourceCheckExistence<Employee>;
+
+  getPrivateEndpointConnection is PrivateEndpointOperations.Read<Employee>;
+  createOrUpdatePrivateEndpointConnection is PrivateEndpointOperations.CreateOrUpdateAsync<Employee>;
+  updatePrivateEndpointConnection is PrivateEndpointOperations.CustomPatchAsync<Employee>;
+  deletePrivateEndpointConnection is PrivateEndpointOperations.DeleteAsync<Employee>;
+  listPrivateEndpointConnections is PrivateEndpointOperations.ListByParent<Employee>;
+}
+
+/** Employee move request */
+model MoveRequest {
+  /** The moving from location */
+  from: string;
+
+  /** The moving to location */
+  to: string;
+}
+
+/** Employee move response */
+model MoveResponse {
+  /** The status of the move */
+  movingStatus: string;
+}
+
+@armResourceOperations
+interface Dependents {
+  get is ArmResourceRead<Dependent>;
+  createOrUpdate is ArmResourceCreateOrReplaceAsync<Dependent>;
+  update is ArmCustomPatchSync<
+    Dependent,
+    Azure.ResourceManager.Foundations.ResourceUpdateModel<Dependent, DependentProperties>
+  >;
+  delete is ArmResourceDeleteSync<Dependent>;
+  list is ArmResourceListByParent<Dependent>;
+  getPrivateEndpointConnection is PrivateEndpointOperations.Read<Dependent>;
+  createOrUpdatePrivateEndpointConnection is PrivateEndpointOperations.CreateOrUpdateAsync<Dependent>;
+  updatePrivateEndpointConnection is PrivateEndpointOperations.CustomPatchAsync<Dependent>;
+  deletePrivateEndpointConnection is PrivateEndpointOperations.DeleteAsync<Dependent>;
+  listPrivateEndpointConnections is PrivateEndpointOperations.ListByParent<Dependent>;
+}
+
+/** An employee dependent */
+@parentResource(Employee)
+model Dependent is ProxyResource<DependentProperties> {
+  ...ResourceNameParameter<Dependent>;
+}
+
+/** Dependent properties */
+model DependentProperties {
+  /** Age of dependent */
+  age: int32;
+
+  /** Gender of dependent */
+  gender: string;
+
+  /** The status of the last operation. */
+  @visibility(Lifecycle.Read)
+  provisioningState?: ProvisioningState;
+}
+`);
+    expectDiagnosticEmpty(diagnostics);
+    const resources = resolveArmResources(program);
+    expect(resources).toBeDefined();
+    expect(resources.resources).toBeDefined();
+    expect(resources.resources).toHaveLength(3);
+    ok(resources.resources);
+    const employee = resources.resources[0];
+    ok(employee);
+    expect(employee).toMatchObject({
+      kind: "Tracked",
+      providerNamespace: "Microsoft.ContosoProviderHub",
+      type: expect.anything(),
+      operations: expect.any(Array),
+    });
+    ok(employee.operations);
+    const subscriptionScope = employee.operations[0];
+    ok(subscriptionScope);
+    checkResolvedOperations(subscriptionScope, {
+      operations: {
+        lifecycle: {},
+        lists: [{ operationGroup: "Employees", name: "listBySubscription", kind: "list" }],
+      },
+      resourceType: {
+        provider: "Microsoft.ContosoProviderHub",
+        types: ["employees"],
+      },
+      resourceInstancePath:
+        "/subscriptions/{subscriptionId}/providers/Microsoft.ContosoProviderHub/employees/{name}",
+    });
+
+    const mainScope = employee.operations[1];
+    ok(mainScope);
+    checkResolvedOperations(mainScope, {
+      operations: {
+        lifecycle: {
+          createOrUpdate: [
+            { operationGroup: "Employees", name: "createOrUpdate", kind: "createOrUpdate" },
+          ],
+          delete: [{ operationGroup: "Employees", name: "delete", kind: "delete" }],
+          read: [{ operationGroup: "Employees", name: "get", kind: "read" }],
+          update: [{ operationGroup: "Employees", name: "update", kind: "update" }],
+        },
+        actions: [{ operationGroup: "Employees", name: "move", kind: "action" }],
+        lists: [{ operationGroup: "Employees", name: "listByResourceGroup", kind: "list" }],
+      },
+      resourceType: {
+        provider: "Microsoft.ContosoProviderHub",
+        types: ["employees"],
+      },
+      resourceInstancePath:
+        "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ContosoProviderHub/employees/{employeeName}",
+    });
+
+    const dependent = resources.resources[2];
+    ok(dependent);
+    expect(dependent).toMatchObject({
+      kind: "Proxy",
+      providerNamespace: "Microsoft.ContosoProviderHub",
+      type: expect.anything(),
+      operations: expect.any(Array),
+    });
+    ok(dependent.operations);
+    expect(dependent.operations).toHaveLength(1);
+    const instanceScope = dependent.operations[0];
+    ok(instanceScope);
+    checkResolvedOperations(instanceScope, {
+      operations: {
+        lifecycle: {
+          createOrUpdate: [
+            { operationGroup: "Dependents", name: "createOrUpdate", kind: "createOrUpdate" },
+          ],
+          delete: [{ operationGroup: "Dependents", name: "delete", kind: "delete" }],
+          read: [{ operationGroup: "Dependents", name: "get", kind: "read" }],
+          update: [{ operationGroup: "Dependents", name: "update", kind: "update" }],
+        },
+        lists: [{ operationGroup: "Dependents", name: "list", kind: "list" }],
+      },
+      resourceType: {
+        provider: "Microsoft.ContosoProviderHub",
+        types: ["employees", "dependents"],
+      },
+      resourceInstancePath:
+        "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ContosoProviderHub/employees/{employeeName}/dependents/{dependentName}",
+    });
+
+    const privateEndpointConnection = resources.resources[1];
+    ok(privateEndpointConnection);
+    ok(privateEndpointConnection.operations);
+    expect(privateEndpointConnection.operations).toHaveLength(2);
+    const privateForEmplInstance = privateEndpointConnection.operations[0];
+    ok(privateForEmplInstance);
+    checkResolvedOperations(privateForEmplInstance, {
+      operations: {
+        lifecycle: {
+          createOrUpdate: [
+            {
+              operationGroup: "Employees",
+              name: "createOrUpdatePrivateEndpointConnection",
+              kind: "createOrUpdate",
+            },
+          ],
+          delete: [
+            {
+              operationGroup: "Employees",
+              name: "deletePrivateEndpointConnection",
+              kind: "delete",
+            },
+          ],
+          read: [
+            { operationGroup: "Employees", name: "getPrivateEndpointConnection", kind: "read" },
+          ],
+          update: [
+            {
+              operationGroup: "Employees",
+              name: "updatePrivateEndpointConnection",
+              kind: "update",
+            },
+          ],
+        },
+        lists: [
+          { operationGroup: "Employees", name: "listPrivateEndpointConnections", kind: "list" },
+        ],
+      },
+      resourceType: {
+        provider: "Microsoft.ContosoProviderHub",
+        types: ["employees", "privateEndpointConnections"],
+      },
+      resourceInstancePath:
+        "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ContosoProviderHub/employees/{employeeName}/privateEndpointConnections/{privateEndpointConnectionName}",
+    });
+
+    const privateForDepInstance = privateEndpointConnection.operations[1];
+    ok(privateForDepInstance);
+
+    checkResolvedOperations(privateForDepInstance, {
+      operations: {
+        lifecycle: {
+          createOrUpdate: [
+            {
+              operationGroup: "Dependents",
+              name: "createOrUpdatePrivateEndpointConnection",
+              kind: "createOrUpdate",
+            },
+          ],
+          delete: [
+            {
+              operationGroup: "Dependents",
+              name: "deletePrivateEndpointConnection",
+              kind: "delete",
+            },
+          ],
+          read: [
+            { operationGroup: "Dependents", name: "getPrivateEndpointConnection", kind: "read" },
+          ],
+          update: [
+            {
+              operationGroup: "Dependents",
+              name: "updatePrivateEndpointConnection",
+              kind: "update",
+            },
+          ],
+        },
+        lists: [
+          { operationGroup: "Dependents", name: "listPrivateEndpointConnections", kind: "list" },
+        ],
+      },
+      resourceType: {
+        provider: "Microsoft.ContosoProviderHub",
+        types: ["employees", "dependents", "privateEndpointConnections"],
+      },
+      resourceInstancePath:
+        "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ContosoProviderHub/employees/{employeeName}/dependents/{dependentName}/privateEndpointConnections/{privateEndpointConnectionName}",
+    });
+
+    checkArmOperationsHas(resources.unassociatedOperations, [
+      { operationGroup: "Employees", name: "checkExistence", kind: "other" },
+      { operationGroup: "Operations", name: "list", kind: "other" },
+    ]);
+  });
 });