+++ packages/typespec-autorest/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log - @azure-tools/typespec-autorest
 
+## 0.59.1
+
+### Bug Fixes
+
+- [#3173](https://github.com/Azure/typespec-azure/pull/3173) Inline azureLocation
+- [#3147](https://github.com/Azure/typespec-azure/pull/3147) Add support for x-ms-external through armExternalResource decorator
+- [#3147](https://github.com/Azure/typespec-azure/pull/3147) Add support for x-ms-azure-resource extension for custom resources
+
+
+## 0.59.0
+
+### Features
+
+- [#3126](https://github.com/Azure/typespec-azure/pull/3126) Add multi-level discriminator support to typespec-autorest
+- [#3125](https://github.com/Azure/typespec-azure/pull/3125) Add  uniqueItems support to Azure.Core and typespec-autorest
+
+### Bump dependencies
+
+- [#3029](https://github.com/Azure/typespec-azure/pull/3029) Upgrade dependencies
+
+
 ## 0.58.1
 
 ### Bug Fixes